+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Release 1.0.0
 
+* [knife-windows #293](https://github.com/chef/knife-windows/issues/293) bootstrap fails on windows 2012 when cmd extensions are disabled in the registry
 * [knife-windows #281](https://github.com/chef/knife-windows/pull/281) Prevent unencrypted negotiate auth, automatically prefix local usernames with '.' for negotiate
 * [knife-windows #275](https://github.com/chef/knife-windows/pull/275) Added bootstrap\_install\_command option in parity with knife bootstrap
 * [knife-windows #240](https://github.com/chef/knife-windows/pull/240) Change kerberos keytab short option to -T to resolve conflict